Extract workspace symbol domain types
