+++ crates/perl-workspace-index/src/workspace/workspace_index.rs
@@ -66,8 +66,10 @@ use crate::ast::{Node, NodeKind};
 use crate::document_store::{Document, DocumentStore};
 use crate::position::{Position, Range};
 use parking_lot::{Mutex, RwLock};
-use perl_position_tracking::{WireLocation, WirePosition, WireRange};
-use serde::{Deserialize, Serialize};
+pub use perl_workspace_symbol_types::{
+    Location, LspWorkspaceSymbol, ReferenceKind, SymKind, SymbolKey, SymbolKind, SymbolReference,
+    VarKind, WorkspaceSymbol, normalize_var, sigil_to_var_kind,
+};
 use std::collections::hash_map::DefaultHasher;
 use std::collections::{HashMap, HashSet};
 use std::hash::{Hash, Hasher};
@@ -1354,182 +1356,6 @@ impl IndexCoordinator {
 impl Default for IndexCoordinator {
     fn default() -> Self {
         Self::new()
-    }
-}
-
-// ============================================================================
-// Symbol Indexing Types
-// ============================================================================
-
-#[derive(Clone, Copy, Debug, Eq, PartialEq, Hash)]
-/// Symbol kinds for cross-file indexing during Index/Navigate workflows.
-pub enum SymKind {
-    /// Variable symbol ($, @, or % sigil)
-    Var,
-    /// Subroutine definition (sub foo)
-    Sub,
-    /// Package declaration (package Foo)
-    Pack,
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, Hash)]
-/// A normalized symbol key for cross-file lookups in Index/Navigate workflows.
-pub struct SymbolKey {
-    /// Package name containing this symbol
-    pub pkg: Arc<str>,
-    /// Bare name without sigil prefix
-    pub name: Arc<str>,
-    /// Variable sigil ($, @, or %) if applicable
-    pub sigil: Option<char>,
-    /// Kind of symbol (variable, subroutine, package)
-    pub kind: SymKind,
-}
-
-/// Normalize a Perl variable name for Index/Analyze workflows.
-///
-/// Extracts an optional sigil and bare name for consistent symbol indexing.
-///
-/// # Arguments
-///
-/// * `name` - Variable name from Perl source, with or without sigil.
-///
-/// # Returns
-///
-/// `(sigil, name)` tuple with the optional sigil and normalized identifier.
-///
-/// # Examples
-///
-/// ```rust,ignore
-/// use perl_parser::workspace_index::normalize_var;
-///
-/// assert_eq!(normalize_var("$count"), (Some('$'), "count"));
-/// assert_eq!(normalize_var("process_emails"), (None, "process_emails"));
-/// ```
-pub fn normalize_var(name: &str) -> (Option<char>, &str) {
-    if name.is_empty() {
-        return (None, "");
-    }
-
-    // Safe: we've checked that name is not empty
-    let Some(first_char) = name.chars().next() else {
-        return (None, name); // Should never happen but handle gracefully
-    };
-    match first_char {
-        '$' | '@' | '%' => {
-            if name.len() > 1 {
-                (Some(first_char), &name[1..])
-            } else {
-                (Some(first_char), "")
-            }
-        }
-        _ => (None, name),
-    }
-}
-
-// Using lsp_types for Position and Range
-
-#[derive(Debug, Clone)]
-/// Internal location type used during Navigate/Analyze workflows.
-pub struct Location {
-    /// File URI where the symbol is located
-    pub uri: String,
-    /// Line and character range within the file
-    pub range: Range,
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize)]
-/// A symbol in the workspace for Index/Navigate workflows.
-pub struct WorkspaceSymbol {
-    /// Symbol name without package qualification
-    pub name: String,
-    /// Type of symbol (subroutine, variable, package, etc.)
-    pub kind: SymbolKind,
-    /// File URI where the symbol is defined
-    pub uri: String,
-    /// Line and character range of the symbol definition
-    pub range: Range,
-    /// Fully qualified name including package (e.g., "Package::function")
-    pub qualified_name: Option<String>,
-    /// POD documentation associated with the symbol
-    pub documentation: Option<String>,
-    /// Name of the containing package or class
-    pub container_name: Option<String>,
-    /// Whether this symbol has a body (false for forward declarations)
-    #[serde(default = "default_has_body")]
-    pub has_body: bool,
-}
-
-fn default_has_body() -> bool {
-    true
-}
-
-// Re-export the unified symbol types from perl-symbol-types
-/// Symbol kind enums used during Index/Analyze workflows.
-pub use perl_symbol_types::{SymbolKind, VarKind};
-
-/// Helper function to convert sigil to VarKind
-fn sigil_to_var_kind(sigil: &str) -> VarKind {
-    match sigil {
-        "@" => VarKind::Array,
-        "%" => VarKind::Hash,
-        _ => VarKind::Scalar, // Default to scalar for $ and unknown
-    }
-}
-
-#[derive(Debug, Clone)]
-/// Reference to a symbol for Navigate/Analyze workflows.
-pub struct SymbolReference {
-    /// File URI where the reference occurs
-    pub uri: String,
-    /// Line and character range of the reference
-    pub range: Range,
-    /// How the symbol is being referenced (definition, usage, etc.)
-    pub kind: ReferenceKind,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-/// Classification of how a symbol is referenced in Navigate/Analyze workflows.
-pub enum ReferenceKind {
-    /// Symbol definition site (sub declaration, variable declaration)
-    Definition,
-    /// General usage of the symbol (function call, method call)
-    Usage,
-    /// Import via use statement
-    Import,
-    /// Variable read access
-    Read,
-    /// Variable write access (assignment target)
-    Write,
-}
-
-#[derive(Debug, Serialize)]
-#[serde(rename_all = "camelCase")]
-/// LSP-compliant workspace symbol for wire format in Navigate/Analyze workflows.
-pub struct LspWorkspaceSymbol {
-    /// Symbol name as displayed to the user
-    pub name: String,
-    /// LSP symbol kind number (see lsp_types::SymbolKind)
-    pub kind: u32,
-    /// Location of the symbol definition
-    pub location: WireLocation,
-    /// Name of the containing symbol (package, class)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub container_name: Option<String>,
-}
-
-impl From<&WorkspaceSymbol> for LspWorkspaceSymbol {
-    fn from(sym: &WorkspaceSymbol) -> Self {
-        let range = WireRange {
-            start: WirePosition { line: sym.range.start.line, character: sym.range.start.column },
-            end: WirePosition { line: sym.range.end.line, character: sym.range.end.column },
-        };
-
-        Self {
-            name: sym.name.clone(),
-            kind: sym.kind.to_lsp_kind(),
-            location: WireLocation { uri: sym.uri.clone(), range },
-            container_name: sym.container_name.clone(),
-        }
     }
 }
